requirements.txt
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
requests>=2.31.0
openai>=1.0.0
twilio>=8.10.0
streamlit>=1.28.0
plotly>=5.18.0
cryptography>=41.0.0
keyring>=23.14.0
schedule>=1.2.0
rich>=13.7.0
beautifulsoup4>=4.12.0
gitpython>=3.1.0
pyyaml>=6.0.0
duckduckgo-search>=7.0.0